fixed a potential memory leaking - if an exception was thrown, the feeds file might not be deleted

+++ FeedHandler.php
@@ -152,7 +152,10 @@ abstract class FeedHandler
 
   public function deleteFeed()
   {
-    unlink($this->feedFilepath);
+    if (is_file($this->feedFilepath))
+    {
+      unlink($this->feedFilepath);
+    }
   }
 
   public function setItemArray($itemArray)
